Add empty AnalysisGrid class for tomorrow
