+++ Environment_oM/Environment_oM.csproj
@@ -41,6 +41,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Analysis\AnalysisGrid.cs" />
     <Compile Include="Climate\IClimateObject.cs" />
     <Compile Include="Climate\Location.cs" />
     <Compile Include="Climate\SpaceTime.cs" />
@@ -51,7 +52,7 @@
     <Compile Include="Elements\Enums\OpeningType.cs" />
     <Compile Include="Elements\Enums\PanelType.cs" />
     <Compile Include="Elements\Enums\SpaceType.cs" />
-    <Compile Include="Elements\Node.cs" />
+    <Compile Include="Analysis\Node.cs" />
     <Compile Include="Elements\Opening.cs" />
     <Compile Include="Elements\Panel.cs" />
     <Compile Include="Elements\Space.cs" />
